refactor: separate forward and reverse mixins to separate file
